pyproject.toml
[project]
name = "learn-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
]
